[fix] aligned pixel perfection workable, but not done yet

+++ src/ui/frame/frame_widget.cpp
@@ -136,12 +136,12 @@ void ekg::ui::frame::on_event(ekg::io::stage stage) {
             this->descriptor.rect.x = new_rect.x - this->p_parent_rect->x;
             this->descriptor.rect.y = new_rect.y - this->p_parent_rect->y;
           } else {
-            this->descriptor.rect.x = new_rect.x;
-            this->descriptor.rect.y = new_rect.y;
+            this->descriptor.rect.x = (int32_t) new_rect.x;
+            this->descriptor.rect.y = (int32_t) new_rect.y;
           }
 
-          this->descriptor.rect.w = new_rect.w;
-          this->descriptor.rect.h = new_rect.h;
+          this->descriptor.rect.w = (int32_t) new_rect.w;
+          this->descriptor.rect.h = (int32_t) new_rect.h;
 
           if (this->target_dock_resize != ekg::dock::none) {
             this->states.is_targeting_absolute_parent = true;
